backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-dateutil==2.8.2
